尝试优化 pwa install 实现
